Move StoredMessage class to its own python module
