+++ bot/action/extra/messages/__init__.py
@@ -3,14 +3,14 @@ import json
 
 from bot.action.chatsettings import ChatSettings
 from bot.action.core.action import Action
-from bot.action.core.command import UnderscoredCommandBuilder
 from bot.action.core.command.usagemessage import CommandUsageMessage
 from bot.action.extra.messages import analyzer
 from bot.action.extra.messages.mapper import StoredMessageMapper
+from bot.action.extra.messages.stored_message import StoredMessage
 from bot.action.userinfo import UserStorageHandler
-from bot.action.util.format import UserFormatter, DateFormatter
+from bot.action.util.format import UserFormatter
 from bot.action.util.textformat import FormattedText
-from bot.api.domain import Message, ApiObject
+from bot.api.domain import Message
 
 MIN_MESSAGES_TO_KEEP = 5000
 MAX_MESSAGES_TO_KEEP = 15000
@@ -217,33 +217,6 @@ class ListMessageAction(Action):
                 .bold("⚠️ Opt-out override is currently enabled on this chat ⚠️").newline()\
                 .normal("Opt-out list is not in effect here while override is enabled.")
         return response.build_message()
-
-
-class StoredMessage:
-    def __init__(self, message_id, message, *edited_messages):
-        self.message_id = message_id
-        self.message = message
-        self.edited_messages = edited_messages
-
-    @property
-    def user_id(self):
-        return self.message.from_
-
-    def printable_info(self, event, user_storage_handler):
-        show_command = UnderscoredCommandBuilder.build_command(event.command, self.message_id)
-        return analyzer.get_short_info(user_storage_handler, self, show_command)
-
-    def printable_full_message(self, user_storage_handler):
-        return analyzer.get_full_content(user_storage_handler, self)
-
-    @staticmethod
-    def deserialize(message_id, data):
-        messages = []
-        for line in data.splitlines():
-            message_data = json.loads(line)
-            message = ApiObject.wrap_api_object(message_data)
-            messages.append(message)
-        return StoredMessage(message_id, *messages)
 
 
 class MessageList:
